Move ast_vector to pool
