+++ src/ast_method.c
@@ -1,5 +1,4 @@
 #include "ast_method.h"
-#include "ast/ast_vector.h"
 
 void ast_initialize_pool(void) {
   static bool is_first = true;
